Refactor asakusafw-organizer plug-in.
